requirements.txt
httpx
pandas
pypdf
pyyaml
tqdm
openpyxl
httpx[http2]